requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
airportsdata==20250909
altair==6.0.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
argcomplete==3.6.3
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blake3==1.0.8
blinker==1.9.0
boto3==1.40.73
botocore==1.40.73
brotli==1.2.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.1
comm==0.2.3
compressed-tensors==0.9.2
contourpy==1.3.2
cryptography==46.0.3
cupy-cuda12x==13.6.0
cycler==0.12.1
dacite==1.9.2
decorator==5.2.1
decord==0.6.0
Deprecated==1.3.1
depyf==0.18.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@4841e70ee48da72c32304f9ebf98138c2a70048d
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
effdet==0.4.1
einops==0.8.1
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.120.3
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastrlock==0.8.3
ffmpy==0.6.4
fiftyone==1.10.0
fiftyone-brain==0.21.4
fiftyone_db==1.4.0
filelock==3.19.1
filetype==1.2.0
flash_attn==2.8.3
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.1.1
future==1.0.0
fvcore==0.1.5.post20221221
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.45
glob2==0.7
google-api-core==1.34.1
google-auth==2.43.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.72.0
gradio==4.44.1
gradio_bbox_annotator==0.1.1
gradio_client==1.3.0
gradio_image_prompter==0.1.0
graphql-core==3.2.7
grpcio==1.76.0
grpcio-status==1.48.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanize==4.14.0
hydra-core==1.3.2
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.7
ImageIO==2.37.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflate64==1.0.3
interegular==0.3.3
iopath==0.1.10
ipython==8.37.0
ipywidgets==8.1.7
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyterlab_widgets==3.0.15
kagglehub==0.3.13
kiwisolver==1.4.9
lark==1.2.2
layoutparser==0.3.4
lazy_loader==0.4
lia-web==0.2.3
llguidance==0.7.30
llvmlite==0.43.0
lm-format-enforcer==0.10.12
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapipe==0.10.21
mistral_common==1.8.5
ml_dtypes==0.5.3
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
multivolumefile==0.2.3
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
numba==0.60.0
numpy==1.26.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==2.6.1
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
orjson==3.11.4
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.3.3
paramiko==3.5.1
parso==0.8.5
partial-json-parser==0.2.1.1.post6
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect==4.9.0
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
plotly==6.4.0
polars==1.35.1
polars-runtime-32==1.35.1
portalocker==3.2.0
pprintpp==0.4.0
priority==2.0.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==3.20.3
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py7zr==1.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.6
pycocotools==2.0.10
pycountry==24.6.1
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.10.6
pydantic-extra-types==2.10.6
pydantic_core==2.27.2
pydash==8.0.5
pydeck==0.9.1
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
pymongo==4.9.2
PyNaCl==1.6.1
pyparsing==3.2.5
pypdfium2==5.0.0
pyppmd==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
pyzstd==0.18.0
qwen-vl-utils==0.0.14
rarfile==4.2
ray==2.51.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
retrying==1.4.2
-e git+https://github.com/IDEA-Research/Rex-Omni.git@0d5fd7832c5cdca048c0daa64e2bcec49ee79161#egg=rex_omni
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.2
roboflow==1.2.11
rpds-py==0.28.0
rsa==4.9.1
rtree==1.4.1
ruff==0.14.2
s3transfer==0.14.0
safetensors==0.6.2
sam2==1.1.0
-e git+https://github.com/facebookresearch/sam3.git@b26a5f330e05d321afb39d01d3d4881f258f65ff#egg=sam3
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.43.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.3
soupsieve==2.8
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.49.1
strawberry-graphql==0.285.0
streamlit==1.52.1
sympy==1.13.1
tabulate==0.9.0
taskgroup==0.2.2
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
timm==1.0.22
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
tomlkit==0.12.0
torch==2.6.0+cu126
torchaudio==2.6.0
torchvision==0.21.0+cu126
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.2.0
typer==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.223
ultralytics-thop==2.0.18
universal-analytics-python3==1.1.1
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.8.2
voxel51-eta==0.15.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==2.0.1
wsproto==1.3.1
xformers==0.0.29.post2
xgrammar==0.1.16
xmltodict==1.0.2
yacs==0.1.8
yarl==1.22.0
zipp==3.23.0
